requirements.txt
python-dotenv>=0.19.0
telethon>=1.24.0
cryptg>=0.2.post4
sqlalchemy==1.3.20
psycopg2-binary>=2.9.1
asyncio>=3.4.3
importlib-metadata>=4.8.1
typing-extensions>=3.10.0
